fix(internal): clear -Wsign-conversion in blockbinary + blocksignificand (#1280)
Sub-issue of Epic #1279. All fixes are value-preserving; each casts a
value already proven non-negative / in range at the site.

blockbinary.hpp:
- operator=(long long): `rhs & storageMask` promoted the signed rhs to
  uint64_t (storageMask is uint64_t); cast rhs explicitly -- the low
  bits kept by the mask are identical in two's complement
- operator<<= / operator>>=: same shift-path idiom as the integer fix
  (#1283) -- int bitsToShift / blockShift mixing with unsigned
  bitsInBlock / nbits / MSU in the block-shift index arithmetic;
  bitsToShift is > 0 and < bitsInBlock/nbits past the guards, so the
  casts are exact. Upper-block mask index cast into the shared
  bit_high_mask<bt> helper. Moved one over-120-column trailing comment
  onto its own line.

blocksignificand.hpp: identical shift-path idiom, same treatment
(including the same trailing-comment reflow).

Verified -Wsign-conversion clean for both headers on gcc AND clang under
a multi-config posit/cfloat instantiation; no new -Wconversion introduced.
blockbinary shift_left/shift_right + addition/multiplication/division/
rounding and blocksignificand addition/multiplication/rounding
regressions PASS on both compilers; multi-config posit/cfloat arithmetic
smoke runs clean.

Relates to #1279
Resolves #1280

+++ include/sw/universal/internal/blocksignificand/blocksignificand.hpp
@@ -328,17 +328,17 @@ public:
 				_block[i] = bt(0);
 			}
 			// adjust the shift
-			bitsToShift -= static_cast<int>(blockShift * bitsInBlock);
+			bitsToShift -= static_cast<int>(static_cast<unsigned>(blockShift) * bitsInBlock);
 			if (bitsToShift == 0) return *this;
 		}
 		if constexpr (MSU > 0) {
 			// construct the mask for the upper bits in the block that need to move to the higher word
-			bt mask = bit_high_mask<bt>(bitsToShift, bitsInBlock);
+			bt mask = bit_high_mask<bt>(static_cast<unsigned>(bitsToShift), bitsInBlock);
 			for (unsigned i = MSU; i > 0; --i) {
 				_block[i] <<= bitsToShift;
 				// mix in the bits from the right
 				bt bits = bt(mask & _block[i - 1]);
-				_block[i] |= (bits >> (bitsInBlock - bitsToShift));
+				_block[i] |= (bits >> (bitsInBlock - static_cast<unsigned>(bitsToShift)));
 			}
 		}
 		_block[0] <<= bitsToShift;
@@ -354,7 +354,7 @@ public:
 
 		unsigned blockShift = 0;
 		if (bitsToShift >= static_cast<int>(bitsInBlock)) {
-			blockShift = bitsToShift / bitsInBlock;
+			blockShift = static_cast<unsigned>(bitsToShift) / bitsInBlock;
 			if (MSU >= blockShift) {
 				// shift by blocks
 				for (unsigned i = 0; i <= MSU - blockShift; ++i) {
@@ -366,7 +366,7 @@ public:
 			if (bitsToShift == 0) {
 				// clean up the blocks we have shifted clean
 				bitsToShift += static_cast<int>(blockShift * bitsInBlock);
-				for (unsigned i = nbits - bitsToShift; i < nbits; ++i) {
+				for (unsigned i = nbits - static_cast<unsigned>(bitsToShift); i < nbits; ++i) {
 					this->setbit(i, false); // reset
 				}
 
@@ -375,19 +375,20 @@ public:
 		}
 		if constexpr (MSU > 0) {
 			bt mask = ALL_ONES;
-			mask >>= (bitsInBlock - bitsToShift); // this is a mask for the lower bits in the block that need to move to the lower word
+			// mask for the lower bits in the block that need to move to the lower word
+			mask >>= (bitsInBlock - static_cast<unsigned>(bitsToShift));
 			for (unsigned i = 0; i < MSU; ++i) {  // TODO: can this be improved? we should not have to work on the upper blocks in case we block shifted
 				_block[i] >>= bitsToShift;
 				// mix in the bits from the left
 				bt bits = bt(mask & _block[i + 1]);
-				_block[i] |= (bits << (bitsInBlock - bitsToShift));
+				_block[i] |= (bits << (bitsInBlock - static_cast<unsigned>(bitsToShift)));
 			}
 		}
 		_block[MSU] >>= bitsToShift;
 
 		// clean up the blocks we have shifted clean
 		bitsToShift += static_cast<int>(blockShift * bitsInBlock);
-		for (unsigned i = nbits - bitsToShift; i < nbits; ++i) {
+		for (unsigned i = nbits - static_cast<unsigned>(bitsToShift); i < nbits; ++i) {
 			this->setbit(i, false); // reset
 		}
